pdf feltöltés, letöltés done, videó feltöltés done, current done

+++ src/main/java/com/example/szakdoga/repository/FilesRepository.java
@@ -3,5 +3,5 @@ package com.example.szakdoga.repository;
 import com.example.szakdoga.model.File;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FilesRepository extends JpaRepository<File, Integer> {
+public interface FilesRepository extends JpaRepository<File, Long> {
 }
